requirements.txt
Django>=3.0, < 3.1
# django-recurrence==1.10.3
-e git+https://github.com/jleclanche/django-recurrence.git@master#egg=django-recurrence
psycopg2==2.8.5
gunicorn==20.0.4
certifi
ofxtools==0.8.21
celery==4.4.6
django-celery-results==1.2.1
django-celery-beat==2.0.0
redis==3.5.3
django-autocomplete-light==3.5.1
pytest-django==4.1.0
